Move all steps to candidate helper
Addressing PR comments
