+++ spec/system/candidate_interface/continuous_applications/submitting/candidate_submitting_application_spec.rb
@@ -3,7 +3,6 @@ require 'rails_helper'
 RSpec.feature 'Candidate submits the application' do
   include CandidateHelper
   include SignInHelper
-  include CandidateSubmissionHelper
 
   scenario 'Candidate with a completed application' do
     given_i_am_signed_in
