+++ spec/system/candidate_interface/continuous_applications/submitting/candidate_tries_to_submit_an_application_for_course_unavailable_spec.rb
@@ -2,7 +2,7 @@ require 'rails_helper'
 
 RSpec.feature 'Candidate tries to submit an application choice when the course is unavailable' do
   include SignInHelper
-  include CandidateSubmissionHelper
+  include CandidateHelper
   before do
     given_i_am_signed_in
     and_i_have_one_application_in_draft
